取得貼文(測試完畢>取得單一帳號貼文資料58篇)

+++ src/main/java/org/example/strategy/GetMediaStrategy.java
@@ -1,6 +1,7 @@
 package org.example.strategy;
 
 import lombok.extern.slf4j.Slf4j;
+import org.example.bean.enumtype.TaskStatusEnum;
 import org.example.entity.LoginAccount;
 import org.example.entity.TaskQueue;
 import org.example.service.InstagramService;
@@ -35,8 +36,9 @@ public class GetMediaStrategy extends TaskStrategyBase implements TaskStrategy {
         log.info("開始執行任務:{} ,帳號:{}", taskQueue.getTaskConfig().getTaskType(), loginAccount);
         //登入、檢查結果並更新登入帳號狀態
         loginAndUpdateAccountStatus(loginAccount);
-        //刪除舊的媒體資料
-        mediaService.deleteOldMediaDataByIgUserId(taskQueue.getIgUser().getId());
+        //初次進行時刪除舊的媒體資料
+        if (TaskStatusEnum.PENDING.equals(taskQueue.getStatus()))
+            mediaService.deleteOldMediaDataByIgUserId(taskQueue.getIgUser().getId());
         //執行爬蟲任務
         performTaskWithAccount(taskQueue);
         //結束任務，依條件判斷更新任務狀態
@@ -90,7 +92,7 @@ public class GetMediaStrategy extends TaskStrategyBase implements TaskStrategy {
     private boolean checkMedia(TaskQueue task) {
         int crawlerAmount = mediaService.countMediaByIgUser(task.getIgUser());
         int dbAmount = task.getIgUser().getMediaCount();
-        log.info("任務:{} ,取追蹤者數量:{},資料庫追蹤者數量:{}", task, dbAmount, crawlerAmount);
+        log.info("任務:{} ,取得貼文數量:{},資料庫貼文數量:{}", task, dbAmount, crawlerAmount);
 
         // 計算當前日期-1年
         LocalDateTime cutoffDate = LocalDateTime.now().minusYears(1);
